move agent stuff to an agent folder
