+++ frontend/src/app/workspace/component/agent/agent-interaction/agent-interaction.component.scss
