+++ frontend/src/app/workspace/component/agent/agent-panel/agent-chat/agent-chat.component.scss
