+++ frontend/src/app/workspace/component/agent/agent-panel/agent-panel.component.scss
