+++ frontend/src/app/workspace/component/agent/agent-panel/agent-registration/agent-registration.component.scss
