+++ frontend/src/app/workspace/component/agent/agent-panel/react-step-detail-modal/react-step-detail-modal.component.ts
@@ -18,7 +18,7 @@
  */
 
 import { Component, Input, Output, EventEmitter } from "@angular/core";
-import { ReActStep } from "../../../service/agent/agent-types";
+import { ReActStep } from "../../../../service/agent/agent-types";
 
 /**
  * Reusable modal component for displaying ReActStep details.
